// src/app/core/auto_api/backend/model/briefTemplate.ts
/**
 * Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EnrollmentPeriod } from './enrollmentPeriod';
import { ProfileBriefTemplate } from './profileBriefTemplate';
import { User } from './user';

export interface BriefTemplate { 
    id?: number;
    enrollmentId?: number;
    isDefault?: number;
    attributeName?: string;
    attributeType?: string;
    attributePlaceholder?: string;
    isRequire?: number;
    creatorId?: number;
    createdAt?: Date;
    updatedAt?: Date;
    creator?: User;
    enrollment?: EnrollmentPeriod;
    profileBriefTemplates?: Array<ProfileBriefTemplate>;
}