// src/app/core/auto_api/backend/model/enrollmentPeriod.ts
/**
 * Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AdmissionProfile } from './admissionProfile';
import { BriefTemplate } from './briefTemplate';
import { ProfileActionLog } from './profileActionLog';
import { ProfileBriefTemplate } from './profileBriefTemplate';
import { User } from './user';

export interface EnrollmentPeriod { 
    id?: number;
    code?: string;
    academicYearPrefix?: string;
    academicYearStart?: number;
    academicYearEnd?: number;
    academicDegree?: string;
    startTime?: Date;
    stepTime?: string;
    doneTime?: Date;
    principalSupervisorId?: number;
    isSendingResumesFreely?: number;
    currentStep?: string;
    createdAt?: Date;
    updatedAt?: Date;
    principalSupervisor?: User;
    admissionProfiles?: Array<AdmissionProfile>;
    briefTemplates?: Array<BriefTemplate>;
    profileActionLogs?: Array<ProfileActionLog>;
    profileBriefTemplates?: Array<ProfileBriefTemplate>;
}